feat: refactor estructura de interfaces y constantes en la gestión de historial médico
